İlac Takibi/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Babel==4.0.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.1.0
folium==0.15.1
geopy==2.4.1
PyJWT==2.8.0
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.18.0
bcrypt==4.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
email-validator==2.1.0.post1 